Set useYoutubeSearch hook
